+++ src/main/java/com/gcf/spring/repository/AdminNoticeRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.gcf.spring.entity.Attachment;
 import com.gcf.spring.entity.Notice;
 
-public interface AdminNoticeRepository extends JpaRepository<Notice, Long>{
-	Optional<Notice> findById(Long id);
+public interface AdminNoticeRepository extends JpaRepository<Notice, Integer>{
+	Optional<Notice> findById(Integer id);
 }
